Extract out mixnet_listener.rs
