+++ service-providers/ip-packet-router/src/lib.rs
@@ -4,11 +4,12 @@
 pub use crate::config::Config;
 pub use ip_packet_router::{IpPacketRouterBuilder, OnStartData};
 
-pub mod config;
 mod constants;
-pub mod error;
 mod ip_packet_router;
 mod mixnet_client;
+mod mixnet_listener;
 mod request_filter;
 mod tun_listener;
 mod util;
+pub mod config;
+pub mod error;
