+++ service-providers/ip-packet-router/src/util/generate_new_ip.rs
@@ -3,7 +3,7 @@ use std::{
     net::{IpAddr, Ipv4Addr},
 };
 
-use crate::{constants::TUN_DEVICE_ADDRESS, ip_packet_router::ConnectedClient};
+use crate::{constants::TUN_DEVICE_ADDRESS, mixnet_listener::ConnectedClient};
 
 // Find an available IP address in self.connected_clients
 // TODO: make this nicer
